remove series.py and move unimplemented stubs
