+++ src/snowflake/snowpark/modin/plugin/extensions/series_overrides.py
@@ -9,8 +9,11 @@ pandas, such as `Series.memory_usage`.
 
 from __future__ import annotations
 
+from typing import Any
+
 import pandas as native_pd
 from modin.pandas import Series
+from pandas._libs.lib import no_default
 
 from snowflake.snowpark.modin import pandas as spd  # noqa: F401
 from snowflake.snowpark.modin.pandas.api.extensions import register_series_accessor
@@ -22,83 +25,225 @@ from snowflake.snowpark.modin.plugin.utils.error_message import series_not_imple
 from snowflake.snowpark.modin.plugin.utils.warning_message import WarningMessage
 from snowflake.snowpark.modin.utils import _inherit_docstrings
 
-# These methods are not implemented by Snowpark pandas and raise errors at the frontend layer
-frontend_not_implemented = [
-    "argmax",
-    "argmin",
-    "argsort",
-    "array",
-    "asfreq",
-    "asof",
-    "at",
-    "at_time",
-    "autocorr",
-    "backfill",
-    "between",
-    "between_time",
-    "bfill",
-    "bool",
-    "clip",
-    "combine",
-    "combine_first",
-    "compare",
-    "divmod",
-    "dot",
-    "droplevel",
-    "ewm",
-    "explode",
-    "factorize",
-    "filter",
-    "hist",
-    "infer_objects",
-    "interpolate",
-    "item",
-    "kurt",
-    "kurtosis",
-    "mode",
-    "nbytes",
-    "nlargest",
-    "nsmallest",
-    "nsmallest",
-    "pipe",
-    "plot",
-    "pop",
-    "prod",
-    "ravel",
-    "reindex_like",
-    "reorder_levels",
-    "repeat",
-    "rdivmod",
-    "searchsorted",
-    "sem",
-    "set_flags",
-    "swapaxes",
-    "swaplevel",
-    "to_clipboard",
-    "to_csv",
-    "to_excel",
-    "to_hdf",
-    "to_json",
-    "to_latex",
-    "to_markdown",
-    "to_period",
-    "to_sql",
-    "to_string",
-    "to_timestamp",
-    "to_xarray",
-    "transform",
-    "truncate",
-    "tz_convert",
-    "tz_localize",
-    "unstack",
-    "view",
-    "xs",
-]
 
-for name in frontend_not_implemented:
-    register_series_accessor(name)(series_not_implemented()(getattr(spd.Series, name)))
+def register_series_not_implemented():
+    def decorator(base_method: Any):
+        func = snowpark_pandas_telemetry_method_decorator(
+            series_not_implemented()(base_method)
+        )
+        register_series_accessor(base_method.__name__)(func)
+        return func
+
+    return decorator
 
 
+# === UNIMPLEMENTED METHODS ===
+# The following methods are not implemented in Snowpark pandas, and must be overridden on the
+# frontend. These methods fall into a few categories:
+# 1. Would work in Snowpark pandas, but we have not tested it.
+# 2. Would work in Snowpark pandas, but requires more SQL queries than we are comfortable with.
+# 3. Requires materialization (usually via a frontend _default_to_pandas call).
+# 4. Performs operations on a native pandas Index object that are nontrivial for Snowpark pandas to manage.
+
+
+@register_series_not_implemented()
+def argsort(self, axis=0, kind="quicksort", order=None):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def transform(self, func, axis=0, *args, **kwargs):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def autocorr(self, lag=1):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def between(self, left, right, inclusive: str = "both"):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def corr(self, other, method="pearson", min_periods=None):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def cov(self, other, min_periods=None, ddof: int | None = 1):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def divmod(self, other, level=None, fill_value=None, axis=0):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def dot(self, other):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def explode(self, ignore_index: bool = False):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def factorize(
+    self, sort=False, na_sentinel=no_default, use_na_sentinel=no_default
+):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def hist(
+    self,
+    by=None,
+    ax=None,
+    grid=True,
+    xlabelsize=None,
+    xrot=None,
+    ylabelsize=None,
+    yrot=None,
+    figsize=None,
+    bins=10,
+    **kwds,
+):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def interpolate(
+    self,
+    method="linear",
+    axis=0,
+    limit=None,
+    inplace=False,
+    limit_direction: str | None = None,
+    limit_area=None,
+    downcast=None,
+    **kwargs,
+):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def item(self):  # noqa: RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def items(self):  # noqa: RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def prod(
+    self,
+    axis=None,
+    skipna=True,
+    level=None,
+    numeric_only=False,
+    min_count=0,
+    **kwargs,
+):
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def ravel(self, order="C"):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def reindex_like(
+    self,
+    other,
+    method=None,
+    copy: bool | None = None,
+    limit=None,
+    tolerance=None,
+) -> Series:
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def reorder_levels(self, order):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def searchsorted(self, value, side="left", sorter=None):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def swaplevel(self, i=-2, j=-1, copy=True):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def to_period(self, freq=None, copy=True):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def to_string(
+    self,
+    buf=None,
+    na_rep="NaN",
+    float_format=None,
+    header=True,
+    index=True,
+    length=False,
+    dtype=False,
+    name=False,
+    max_rows=None,
+    min_rows=None,
+):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def to_timestamp(self, freq=None, how="start", copy=True):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def array(self):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def nbytes(self):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+@register_series_not_implemented()
+def __reduce__(self):  # noqa: PR01, RT01, D200
+    pass  # pragma: no cover
+
+
+# === OVERRIDDEN METHODS ===
+# The below methods have their frontend implementations overridden compared to the version present
+# in series.py. This is usually for one of the following reasons:
+# 1. The underlying QC interface used differs from that of modin. Notably, this applies to aggregate
+#    and binary operations; further work is needed to refactor either our implementation or upstream
+#    modin's implementation.
+# 2. Modin performs extra validation queries that perform extra SQL queries. Some of these are already
+#    fixed on main; see https://github.com/modin-project/modin/issues/7340 for details.
+# 3. Upstream Modin defaults to pandas for some edge cases. Defaulting to pandas at the query compiler
+#    layer is acceptable because we can force the method to raise NotImplementedError, but if a method
+#    defaults at the frontend, Modin raises a warning and performs the operation by coercing the
+#    dataset to a native pandas object. Removing these is tracked by
+#    https://github.com/modin-project/modin/issues/7104
+
+
+# Since Snowpark pandas leaves all data on the warehouse, memory_usage's report of local memory
+# usage isn't meaningful and is set to always return 0.
 @_inherit_docstrings(native_pd.Series.memory_usage, apilink="pandas.Series")
 @register_series_accessor("memory_usage")
 def memory_usage(self, index: bool = True, deep: bool = False) -> int:
@@ -109,6 +254,7 @@ def memory_usage(self, index: bool = True, deep: bool = False) -> int:
     return 0
 
 
+# Snowpark pandas has slightly different type validation from upstream modin.
 @_inherit_docstrings(native_pd.Series.isin, apilink="pandas.Series")
 @register_series_accessor("isin")
 def isin(self, values: set | ListLike) -> Series:
@@ -195,6 +341,7 @@ def isin(self, values: set | ListLike) -> Series:
     return super(Series, self).isin(values)
 
 
+# Snowpark pandas raises a warning before materializing data and passing to `plot`.
 @register_series_accessor("plot")
 @property
 @snowpark_pandas_telemetry_method_decorator
@@ -235,13 +382,6 @@ def plot(
     return self._to_pandas().plot
 
 
-@register_series_accessor("transform")
-@snowpark_pandas_telemetry_method_decorator
-@series_not_implemented()
-def transform(self, func, axis=0, *args, **kwargs):  # noqa: PR01, RT01, D200
-    pass  # pragma: no cover
-
-
 # modin 0.28.1 doesn't define type annotations on properties, so we override this
 # to satisfy test_type_annotations.py
 _old_empty_fget = Series.empty.fget
@@ -253,6 +393,8 @@ def empty(self) -> bool:
     return _old_empty_fget(self)
 
 
+# Upstream modin performs name change and copy operations on binary operators that Snowpark
+# pandas avoids.
 @register_series_accessor("_prepare_inter_op")
 def _prepare_inter_op(self, other):
     # override prevents extra queries from occurring during binary operations
